move test rpc code to crate within openzt for easier testing
