+++ openzt-test-rpc-dll/Cargo.toml
@@ -7,7 +7,7 @@ license.workspace = true
 repository.workspace = true
 
 [dependencies]
-openzt-test-rpc = { path = "../openzt-test-rpc", version = "0.1.0" }
+openzt = { path = "../openzt", version = "0.1.0-alpha", features = ["test-rpc"] }
 
 [target.'cfg(windows)'.dependencies]
 windows =  { version = "0.61.3", features = ["Win32", "Win32_System_Console", "Win32_System_SystemServices"]}
