+++ openzt-test-rpc-dll/src/lib.rs
@@ -9,7 +9,7 @@ use windows::Win32::System::{SystemServices::{DLL_PROCESS_ATTACH, DLL_PROCESS_DE
 extern "system" fn DllMain(_module: u8, reason: u32, _reserved: u8) -> i32 {
     match reason {
         DLL_PROCESS_ATTACH => {
-            openzttestrpclib::init();
+            openztlib::test_rpc::init();
         }
         DLL_PROCESS_DETACH => {}
         DLL_THREAD_ATTACH => {}
