--- openzt-test-rpc/Cargo.toml
@@ -1,35 +1,0 @@
-[package]
-name = "openzt-test-rpc"
-version = "0.1.0"
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[dependencies]
-enable-ansi-support = "0.2.1"
-tarpc = { version = "0.34", features = ["serde-transport", "tcp"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-futures = "0.3"
-tokio = { version = "1", features = ["macros", "net", "rt-multi-thread"] }
-tokio-serde = { version = "0.8", features = ["json"] }
-async-trait = "0.1"
-tracing = "0.1.41"
-tracing-subscriber = "0.3.19"
-retour = { version = "0.3.1", features = ["thiscall-abi"] }
-openzt-detour = { path = "../openzt-detour", version = "0.1.0" }
-openzt-detour-macro = { path = "../openzt-detour-macro", version = "0.1.0" }
-openzt = { path = "../openzt", version = "0.1.0-alpha" }
-paste = "1.0.15"
-
-[target.'cfg(windows)'.dependencies]
-windows =  { version = "0.61.3", features = ["Win32", "Win32_System_Console", "Win32_System_SystemServices"]}
-
-[lib]
-name = "openzttestrpclib"
-crate-type = ["lib"]
-
-[[bin]]
-name = "openzttestrpc_client"
-path = "src/bin.rs"
