+++ openzt/Cargo.toml
@@ -34,10 +34,13 @@ public = "0.2.2"
 enable-ansi-support = "0.2.1"
 itertools = "0.14.0"
 nt-time = "0.10.6"
-tarpc = { version = "0.34", features = ["serde-transport"] }
-serde_json = "1.0"
-futures = "0.3"
-tokio = { version = "1", features = ["macros", "net", "rt-multi-thread"] }
+tarpc = { version = "0.34", features = ["serde-transport", "tcp"], optional = true }
+serde_json = { version = "1.0", optional = true }
+futures = { version = "0.3", optional = true }
+tokio = { version = "1", features = ["macros", "net", "rt-multi-thread"], optional = true }
+tokio-serde = { version = "0.8", features = ["json"], optional = true }
+async-trait = { version = "0.1", optional = true }
+paste = { version = "1.0.15", optional = true }
 
 [target.'cfg(windows)'.dependencies]
 windows = { version = "0.61.3", features = ["Win32", "Win32_System_Console", "Win32_System_SystemServices", "Win32_System_Memory"] }
@@ -46,7 +49,6 @@ windows = { version = "0.61.3", features = ["Win32", "Win32_System_Console", "Wi
 winresource = "0.1"
 
 [dev-dependencies]
-openzt-test-rpc = { path = "../openzt-test-rpc", version = "0.1.0" }
 parking_lot = "0.12"
 proptest = "1.7.0"
 tokio-test = "0.4"
@@ -54,6 +56,11 @@ tokio-test = "0.4"
 [lib]
 name = "openztlib"
 crate-type = ["lib"]
+
+[[bin]]
+name = "openzt_test_rpc_client"
+path = "src/bin/openzt_test_rpc_client.rs"
+required-features = ["test-rpc"]
 
 #[[bin]]
 #name = "test"
@@ -65,3 +72,4 @@ release = []
 ini = []
 capture_ztlog = []
 experimental = []
+test-rpc = ["tarpc", "serde_json", "futures", "tokio", "tokio-serde", "async-trait", "paste"]
